requirements.txt
fastapi
python-telegram-bot
requests
python-dotenv
pydantic
uvicorn